requirements.txt
streamlit==1.41.1
yfinance==0.2.54
altair==5.5.0
pandas==2.2.3
numpy==2.2.2
plotly==6.0.1
google-generativeai==0.8.4
